# app/docs/swagger.yaml
definitions:
...
  models.ChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  models.ChangeUserInfo:
    properties:
      login:
        type: string
      name:
        type: string
      phone:
        type: string
      surname:
        type: string
    type: object
...
  models.EmailChangeComplete:
    properties:
      code:
        type: integer
    type: object
  models.EmailChangeRequest:
    properties:
      email:
        type: string
    type: object
  models.EmailChangeResponse:
    properties:
      access_token:
        type: string
      messages:
        type: string
      refresh_token:
        type: string
      success:
        type: boolean
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.UserInfo:
    properties:
      active:
        type: boolean
      avatar_id:
        type: string
      created:
        type: string
      email:
        type: string
      id:
        type: integer
      login:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: integer
      surname:
        type: string
      updated:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /user/change:
    patch:
      consumes:
      - application/json
      description: Endpoint to change user data
      parameters:
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ChangeUserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Change user data
      tags:
      - User
  /user/change/email:
    post:
      consumes:
      - application/json
      description: Endpoint to change user email
      parameters:
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.EmailChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Change user email
      tags:
      - User
  /user/change/email/submit:
    patch:
      consumes:
      - application/json
      description: Endpoint to complete email change
      parameters:
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.EmailChangeComplete'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmailChangeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Change email complete
      tags:
      - User
  /user/changepass:
    post:
      consumes:
      - application/json
      description: Endpoint to change user password
      parameters:
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Change user password
      tags:
      - User
  /user/info:
    get:
      consumes:
      - application/json
      description: Endpoint to get user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserInfo'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get user info
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"